+++ src/main/java/org/apache/commons/net/ntp/TimeInfo.java
@@ -173,7 +173,7 @@ public class TimeInfo {
             // might be via a broadcast NTP packet...
             if (xmitNtpTime.ntpValue() != 0)
             {
-                _offset = Long.valueOf(xmitTime - _returnTime);
+                _offset = xmitTime - _returnTime;
                 _comments.add("Error: zero orig time -- cannot compute delay");
             } else {
                 _comments.add("Error: zero orig time -- cannot compute delay/offset");
@@ -186,7 +186,7 @@ public class TimeInfo {
             } else {
                 // without receive or xmit time cannot figure out processing time
                 // so delay is simply the network travel time
-                _delay = Long.valueOf(_returnTime - origTime);
+                _delay = _returnTime - origTime;
             }
             // TODO: is offset still valid if rcvNtpTime=0 || xmitNtpTime=0 ???
             // Could always hash origNtpTime (sendTime) but if host doesn't set it
@@ -195,11 +195,11 @@ public class TimeInfo {
             if (rcvNtpTime.ntpValue() != 0)
             {
                 // xmitTime is 0 just use rcv time
-                _offset = Long.valueOf(rcvTime - origTime);
+                _offset = rcvTime - origTime;
             } else if (xmitNtpTime.ntpValue() != 0)
             {
                 // rcvTime is 0 just use xmitTime time
-                _offset = Long.valueOf(xmitTime - _returnTime);
+                _offset = xmitTime - _returnTime;
             }
         } else
         {
@@ -235,12 +235,12 @@ public class TimeInfo {
                     }
                  }
              }
-             _delay = Long.valueOf(delayValue);
+             _delay = delayValue;
             if (origTime > _returnTime) {
                 _comments.add("Error: OrigTime > DestRcvTime");
             }
 
-            _offset = Long.valueOf(((rcvTime - origTime) + (xmitTime - _returnTime)) / 2);
+            _offset = ((rcvTime - origTime) + (xmitTime - _returnTime)) / 2;
         }
     }
